+++ src/nvimage/BlockDXT.h
@@ -116,26 +116,26 @@ namespace nv
 	{
 		union {
 			struct {
-				uint64 alpha0 : 8;	// 8
-				uint64 alpha1 : 8;	// 16
-				uint64 bits0 : 3;	// 3 - 19
-				uint64 bits1 : 3; 	// 6 - 22
-				uint64 bits2 : 3; 	// 9 - 25
-				uint64 bits3 : 3;	// 12 - 28
-				uint64 bits4 : 3;	// 15 - 31
-				uint64 bits5 : 3;	// 18 - 34
-				uint64 bits6 : 3;	// 21 - 37
-				uint64 bits7 : 3;	// 24 - 40
-				uint64 bits8 : 3;	// 27 - 43
-				uint64 bits9 : 3; 	// 30 - 46
-				uint64 bitsA : 3; 	// 33 - 49
-				uint64 bitsB : 3;	// 36 - 52
-				uint64 bitsC : 3;	// 39 - 55
-				uint64 bitsD : 3;	// 42 - 58
-				uint64 bitsE : 3;	// 45 - 61
-				uint64 bitsF : 3;	// 48 - 64
+				unsigned long long alpha0 : 8;	// 8
+				unsigned long long alpha1 : 8;	// 16
+				unsigned long long bits0 : 3;	// 3 - 19
+				unsigned long long bits1 : 3; 	// 6 - 22
+				unsigned long long bits2 : 3; 	// 9 - 25
+				unsigned long long bits3 : 3;	// 12 - 28
+				unsigned long long bits4 : 3;	// 15 - 31
+				unsigned long long bits5 : 3;	// 18 - 34
+				unsigned long long bits6 : 3;	// 21 - 37
+				unsigned long long bits7 : 3;	// 24 - 40
+				unsigned long long bits8 : 3;	// 27 - 43
+				unsigned long long bits9 : 3; 	// 30 - 46
+				unsigned long long bitsA : 3; 	// 33 - 49
+				unsigned long long bitsB : 3;	// 36 - 52
+				unsigned long long bitsC : 3;	// 39 - 55
+				unsigned long long bitsD : 3;	// 42 - 58
+				unsigned long long bitsE : 3;	// 45 - 61
+				unsigned long long bitsF : 3;	// 48 - 64
 			};
-			uint64 u;
+			unsigned long long u;
 		};
 		
 		void evaluatePalette(uint8 alpha[8]) const;
